// src/modules/auth/types/auth-response.type.ts
import { UserStatus } from 'src/generated/prisma/enums';

export type RegisterResponse = {
    id: string;
    username: string;
    email: string;
    firstName: string;
    lastName: string | null;
    photoUrl: string | null;
    status: UserStatus;
    createdAt: Date;
    updatedAt: Date;
};

export type LoginResponse = {
    accessToken: string;
    refreshToken: string;
    user: {
        id: string;
        username: string;
        email: string;
        firstName: string;
        lastName: string | null;
        photoUrl: string | null;
        status: UserStatus;
        createdAt: Date;
        updatedAt: Date;
    };
};